+++ src/resources/views/notification/seatgroup_application/channel.blade.php
@@ -1,4 +1,0 @@
-<blockquote>
-  <p>Not supported</p>
-  <small>Managers may subscribe to new SeAT Group applications individually.</small>
-</blockquote>